int distance

+++ api/src/Entity/Repairer.php
@@ -223,7 +223,7 @@ class Repairer
     public ?\DateTimeInterface $createdAt = null;
 
     #[Groups([self::REPAIRER_COLLECTION_READ])]
-    public ?float $distance = null;
+    public ?int $distance = null;
 
     public function __construct()
     {
